backend/requirements.txt
Flask
langchain
python-dotenv
openai
langchain-community
langchain-openai